rename nlh api file
